Squashed commit of the following:
commit 7ac84597bb1beac123543229f724390b750357af
Author: Junior <losttides@gmail.com>
Date:   Fri Jun 27 20:29:33 2025 -0500

    Protocol Bump and Command ArgumentParser Refactor (#488)

+++ Obsidian/Net/NetworkBuffer.cs
@@ -56,12 +56,12 @@ public partial class NetworkBuffer
     /// <summary>
     /// Get a span of bytes from the current buffer
     /// </summary>
-    public Span<byte> AsSpan() => new(data, (int)offset, (int)size);
+    public Span<byte> AsSpan() => new(data, offset, size);
 
     /// <summary>
     /// Get a span of bytes from the current buffer with the specified size.
     /// </summary>
-    public Span<byte> AsSpan(int size) => new(data, (int)offset, size);
+    public Span<byte> AsSpan(int size) => new(data, offset, size);
 
     public Span<byte> AsSpan(int offset, int? size = null) => new(data, offset, size.HasValue ? size.Value : this.size);
